Modifies product cards color
