+++ burger-queen-api-client/src/requests.http
@@ -1,4 +1,4 @@
-POST https://7487-2806-10ae-10-423f-1d22-ad56-4a4f-9b14.ngrok-free.app/login HTTP/1.1
+POST http://localhost:8080/login HTTP/1.1
 content-type: application/json
 
 {
